ui: move nodeCanvas to new file
Release note: None
